refactor: move playlist selection to own file
